requirements.txt
aiosqlite
anyio
bcrypt
cffi
cryptography
fastapi
h2
h11
hpack
hypercorn
hyperframe
idna
iso8601
jwt
passlib
priority
psycopg2
pycparser
pydantic
python-multipart
pytz
six
sniffio
starlette
tortoise-orm
typing-extensions
uvicorn
toml
wsproto
PyJWT
asyncpg
Jinja2
pypika
fastapi-mail
pypika-tortoise
pandas

